pyproject.toml
[tool.poetry]
name = "template-server"
version = "0.1.0"
description = "Junice App Server Side"
authors = ["test"]
readme = "README.md"
packages = [{ include = "core" }]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110.0"
email-validator = "^2.0.0.post2"
orjson = "^3.9.15"
uvicorn = "^0.29.0"
python-dotenv = "^1.0.0"
asyncpg = "^0.29.0"
passlib = "^1.7.4"
python-jose = "^3.3.0"
fastapi-camelcase = "^2.0.0"
python-multipart = "^0.0.9"
pydantic-settings = "^2.0.3"
bcrypt = "^4.0.1"

[tool.poetry.group.celery.dependencies]
sqlalchemy = "^2.0.28"
psycopg2 = "^2.9.9"
pydantic = "^2.6.3"
pydantic-settings = "^2.0.3"
httpx = "^0.27.0"

[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
httpx = "^0.27.0"
pre-commit = "^3.3.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
